+++ test/data/test_ecmwf.py
@@ -21,7 +21,7 @@ from datetime import datetime, timedelta
 import numpy as np
 import pytest
 
-from earth2studio.data import AIFS, IFS
+from earth2studio.data import AIFS, AIFS_ENS, IFS, IFS_ENS
 
 
 def now6h():
@@ -80,6 +80,55 @@ def test_ifs_fetch(time, lead_time, variable):
 
 @pytest.mark.slow
 @pytest.mark.xfail
+@pytest.mark.timeout(30)
+@pytest.mark.parametrize(
+    "time",
+    [
+        now6h() - timedelta(hours=12),
+        [
+            now6h() - timedelta(days=1),
+            now6h() - timedelta(days=1, hours=6),
+        ],
+    ],
+)
+@pytest.mark.parametrize(
+    "lead_time",
+    [
+        timedelta(hours=0),
+        [
+            timedelta(hours=3),
+            timedelta(hours=6),
+        ],
+    ],
+)
+@pytest.mark.parametrize("variable", ["tcwv", ["sp"]])
+def test_ifs_ens_fetch(time, lead_time, variable):
+    members = [1, 2, 3]
+    ds = IFS_ENS(cache=False, members=members)
+    data = ds(time, lead_time, variable)
+    shape = data.shape
+
+    if isinstance(variable, str):
+        variable = [variable]
+
+    if isinstance(lead_time, timedelta):
+        lead_time = [lead_time]
+
+    if isinstance(time, datetime):
+        time = [time]
+
+    assert shape[0] == len(time)
+    assert shape[1] == len(lead_time)
+    assert shape[2] == len(variable)
+    assert shape[3] == len(members)
+    assert shape[4] == 721
+    assert shape[5] == 1440
+    assert not np.isnan(data.values).any()
+    assert np.array_equal(data.coords["variable"].values, np.array(variable))
+
+
+@pytest.mark.slow
+@pytest.mark.xfail
 @pytest.mark.timeout(60)
 @pytest.mark.parametrize(
     "time",
@@ -122,6 +171,7 @@ def test_ifs_cache(time, lead_time, variable, cache):
 
 
 @pytest.mark.timeout(30)
+@pytest.mark.parametrize("data_source", [IFS, IFS_ENS])
 @pytest.mark.parametrize(
     "time",
     [
@@ -132,13 +182,14 @@ def test_ifs_cache(time, lead_time, variable, cache):
 )
 @pytest.mark.parametrize("lead_time", [timedelta(hours=0)])
 @pytest.mark.parametrize("variable", ["msl"])
-def test_ifs_time_available(time, lead_time, variable):
+def test_ifs_time_available(data_source, time, lead_time, variable):
     with pytest.raises(ValueError):
-        ds = IFS(source="ecmwf")
+        ds = data_source(source="ecmwf")
         ds(time, lead_time, variable)
 
 
 @pytest.mark.timeout(30)
+@pytest.mark.parametrize("data_source", [IFS, IFS_ENS])
 @pytest.mark.parametrize("time", [now6h() - timedelta(days=2)])
 @pytest.mark.parametrize("lead_time", [timedelta(hours=0)])
 @pytest.mark.parametrize(
@@ -150,13 +201,14 @@ def test_ifs_time_available(time, lead_time, variable):
     ],
 )
 @pytest.mark.parametrize("variable", ["msl"])
-def test_ifs_leadtime_available(time, lead_time, variable):
+def test_ifs_leadtime_available(data_source, time, lead_time, variable):
     with pytest.raises(ValueError):
-        ds = IFS()
+        ds = data_source()
         ds(time, lead_time, variable)
 
 
 @pytest.mark.timeout(30)
+@pytest.mark.parametrize("data_source", [AIFS, AIFS_ENS])
 @pytest.mark.parametrize(
     "time",
     [
@@ -167,13 +219,14 @@ def test_ifs_leadtime_available(time, lead_time, variable):
 )
 @pytest.mark.parametrize("lead_time", [timedelta(hours=0)])
 @pytest.mark.parametrize("variable", ["msl"])
-def test_aifs_time_available(time, lead_time, variable):
+def test_aifs_time_available(data_source, time, lead_time, variable):
     with pytest.raises(ValueError):
-        ds = AIFS(source="ecmwf")
+        ds = data_source(source="ecmwf")
         ds(time, lead_time, variable)
 
 
 @pytest.mark.timeout(30)
+@pytest.mark.parametrize("data_source", [AIFS, AIFS_ENS])
 @pytest.mark.parametrize("time", [now6h() - timedelta(days=2)])
 @pytest.mark.parametrize("lead_time", [timedelta(hours=0)])
 @pytest.mark.parametrize(
@@ -185,7 +238,7 @@ def test_aifs_time_available(time, lead_time, variable):
     ],
 )
 @pytest.mark.parametrize("variable", ["msl"])
-def test_aifs_leadtime_available(time, lead_time, variable):
+def test_aifs_leadtime_available(data_source, time, lead_time, variable):
     with pytest.raises(ValueError):
-        ds = AIFS()
+        ds = data_source()
         ds(time, lead_time, variable)